backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pinecone-client
langchain
langchain-openai
langchain-pinecone
langchain-text-splitters
langchain-core
langchain-community
PyMuPDF==1.23.8
pymupdf-fonts==1.0.5
pydantic
uuid